refactor(shipping): Phase 3 IShipmentStore to Domain/Abstractions/
- Move IShipmentStore + companion records (CreateShipmentLine,
  CreateShipmentsResult, ShipmentListFilters) to Domain/Abstractions/
  under namespace Shipping.Service.Domain.Abstractions.
- Update 5 consumer usings (endpoints, carrier polling, 3 handlers);
  add explicit using to ShippingContext + EntityFrameworkExtensions
  since Infrastructure/Data namespace is now empty.
- ShippingContext still implements the interface; split lands in Phase 4.
