+++ shipping-microservice/Shipping.Service/Infrastructure/Data/EntityFramework/ShippingContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using Shipping.Service.Domain;
+using Shipping.Service.Domain.Abstractions;
 
 namespace Shipping.Service.Infrastructure.Data.EntityFramework;
