+++ shipping-microservice/Shipping.Service/IntegrationEvents/EventHandlers/CreateShipmentCommandHandler.cs
@@ -3,7 +3,7 @@ using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Shipping.Service.Contracts.Integration;
-using Shipping.Service.Infrastructure.Data;
+using Shipping.Service.Domain.Abstractions;
 
 namespace Shipping.Service.IntegrationEvents.EventHandlers;
