+++ shipping-microservice/Shipping.Service/IntegrationEvents/EventHandlers/OrderConfirmedEventHandler.cs
@@ -1,7 +1,7 @@
 using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using ECommerce.Shared.Infrastructure.Outbox;
 using Shipping.Service.Contracts.Integration;
-using Shipping.Service.Infrastructure.Data;
+using Shipping.Service.Domain.Abstractions;
 
 namespace Shipping.Service.IntegrationEvents.EventHandlers;
